repository: Implement multi transactions on repository.
* Repository is now safe to be used by concurrent callers for multiple
transactions while keeping them isolated.

* This is achieved by implementing a sort of CoW repository instantiated
when you start a Writing transaction.
