+++ snapshot/store.go
@@ -15,13 +15,22 @@ var (
 
 // RepositoryStore implements btree.Storer
 type SnapshotStore[K any, V any] struct {
-	readonly bool
 	blobtype resources.Type
-	snap     *Snapshot
+
+	// Those two fields are mutually exclusive, only one can be set at a time.
+	snapBuilder *Builder
+	snapReader  *Snapshot
 }
 
 func (s *SnapshotStore[K, V]) Get(sum objects.MAC) (*btree.Node[K, objects.MAC, V], error) {
-	bytes, err := s.snap.repository.GetBlobBytes(s.blobtype, sum)
+	var bytes []byte
+	var err error
+	if s.snapBuilder != nil {
+		bytes, err = s.snapBuilder.repository.GetBlobBytes(s.blobtype, sum)
+	} else {
+		bytes, err = s.snapReader.repository.GetBlobBytes(s.blobtype, sum)
+	}
+
 	if err != nil {
 		return nil, err
 	}
@@ -35,7 +44,7 @@ func (s *SnapshotStore[K, V]) Update(sum objects.MAC, node *btree.Node[K, object
 }
 
 func (s *SnapshotStore[K, V]) Put(node *btree.Node[K, objects.MAC, V]) (objects.MAC, error) {
-	if s.readonly {
+	if s.snapBuilder == nil {
 		return objects.MAC{}, ErrReadOnly
 	}
 
@@ -44,17 +53,16 @@ func (s *SnapshotStore[K, V]) Put(node *btree.Node[K, objects.MAC, V]) (objects.
 		return objects.MAC{}, err
 	}
 
-	mac := s.snap.repository.ComputeMAC(bytes)
-	return mac, s.snap.repository.PutBlobIfNotExists(s.blobtype, mac, bytes)
+	mac := s.snapBuilder.repository.ComputeMAC(bytes)
+	return mac, s.snapBuilder.repository.PutBlobIfNotExists(s.blobtype, mac, bytes)
 }
 
 // persistIndex saves a btree[K, P, V] index to the snapshot.  The
 // pointer type P is converted to a MAC.
-func persistIndex[K any, P comparable, VA, VB any](snap *Snapshot, tree *btree.BTree[K, P, VA], rootres, noderes resources.Type, conv func(VA) (VB, error)) (mac objects.MAC, err error) {
+func persistIndex[K any, P comparable, VA, VB any](snap *Builder, tree *btree.BTree[K, P, VA], rootres, noderes resources.Type, conv func(VA) (VB, error)) (mac objects.MAC, err error) {
 	root, err := btree.Persist(tree, &SnapshotStore[K, VB]{
-		readonly: false,
-		blobtype: noderes,
-		snap:     snap,
+		blobtype:    noderes,
+		snapBuilder: snap,
 	}, conv)
 	if err != nil {
 		return
@@ -72,7 +80,7 @@ func persistIndex[K any, P comparable, VA, VB any](snap *Snapshot, tree *btree.B
 	return mac, snap.repository.PutBlobIfNotExists(rootres, mac, bytes)
 }
 
-func persistMACIndex[K any, P comparable](snap *Snapshot, tree *btree.BTree[K, P, []byte], rootres, noderes, entryres resources.Type) (objects.MAC, error) {
+func persistMACIndex[K any, P comparable](snap *Builder, tree *btree.BTree[K, P, []byte], rootres, noderes, entryres resources.Type) (objects.MAC, error) {
 	return persistIndex(snap, tree, rootres, noderes,
 		func(data []byte) (objects.MAC, error) {
 			mac := snap.repository.ComputeMAC(data)
